# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API Foresy
  version: v1
  description: Documentation de l'API Foresy
paths:
  "/api/v1/auth/login":
    post:
      summary: Test Redis unavailable - fail closed (HTTP 429)
      tags:
        - Authentication
      parameters: []
      responses:
        "200":
          description: user authenticated
        "401":
          description: handles missing X-Forwarded-For by using remote_ip
        "403":
          description: inactive user
        "429":
          description: fail closed when Redis is down
          content:
            application/json:
              examples:
                example_0:
                  value:
                    rate_limit_exceeded:
                      summary: Rate limit exceeded (5 requests/minute)
                      value:
                        error: Rate limit exceeded
                        retry_after: 42
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Rate limit exceeded
                  retry_after:
                    type: integer
                    example: 42
                required:
                  - error
                  - retry_after
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
        required: true
    get:
      summary: GET login endpoint - not rate-limited (endpoint does not exist)
      tags:
        - Authentication
      responses:
        "404":
          description: get login endpoint does not exist (not rate-limited)
...
  "/api/v1/auth/refresh":
    post:
      summary: Refreshes authentication token with rate limiting
      tags:
        - Authentication
      parameters: []
      responses:
        "200":
          description: token refreshed
        "401":
          description: token refreshed (under rate limit)
        "429":
          description: rate limit exceeded - too many refresh attempts
          content:
            application/json:
              examples:
                example_0:
                  value:
                    rate_limit_exceeded:
                      summary: Rate limit exceeded (10 requests/minute)
                      value:
                        error: Rate limit exceeded
                        retry_after: 35
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Rate limit exceeded
                  retry_after:
                    type: integer
                    example: 35
                required:
                  - error
                  - retry_after
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
        required: true
...
servers:
  - url: http://{defaultHost}
    description: Serveur local de développement
    variables:
      defaultHost:
        default: localhost:3000
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearer_auth: []